db-service/requirements.txt
# Copyright 2024 Sustain the Brain

# Core
psycopg2-binary==2.9.10
yoyo-migrations==9.0.0

# Test
ipython==8.29.0
